requirements.txt
boto3
pyserial
requests

